+++ ui/appui-react/src/appui-react/layout/widget/widget-content-layout/header/HeaderSearch.tsx
@@ -1,0 +1,95 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+import "./HeaderSearch.scss";
+import * as React from "react";
+
+import { SearchBox } from "@itwin/itwinui-react";
+
+import { SvgClose, SvgSearch } from "@itwin/itwinui-icons-react";
+import type { WidgetContentLayout } from "../WidgetContentLayout.js";
+import type { HeaderIconToolbar } from "./HeaderIconToolbar.js";
+
+type WidgetContentLayoutHeaderProps = React.ComponentProps<
+  typeof WidgetContentLayout.Header
+>;
+type HeaderIconToolbarProps = React.ComponentProps<typeof HeaderIconToolbar>;
+
+interface HeaderSearchProps
+  extends Pick<WidgetContentLayoutHeaderProps, "iconSize" | "onSearch">,
+    Pick<HeaderIconToolbarProps, "searchExpandedState"> {
+  /** State tuple for the search text value and its setter function. */
+  searchState: [string, React.Dispatch<React.SetStateAction<string>>];
+}
+
+/**
+ * A collapsible search box component for widget headers.
+ * Provides an expandable search input with icons that automatically focuses when expanded.
+ * @internal
+ */
+export function HeaderSearch(props: HeaderSearchProps) {
+  const [isExpanded, setIsExpanded] = props.searchExpandedState;
+  const [searchText, setSearchText] = props.searchState;
+  const inputRef = React.useRef<HTMLInputElement | null>(null);
+
+  React.useEffect(() => {
+    if (!isExpanded) return;
+    const el = inputRef.current;
+    if (!el) return;
+    // Focus on next animation frame to allow expand animation / layout to settle.
+    const id = requestAnimationFrame(() => {
+      el.focus();
+      const v = el.value;
+      if (v) el.selectionStart = el.selectionEnd = v.length;
+    });
+    return () => cancelAnimationFrame(id);
+  }, [isExpanded]);
+
+  return (
+    <SearchBox
+      onExpand={() => setIsExpanded(true)}
+      isExpanded={isExpanded}
+      className="nz-widget-widgetContentLayout-header-headerSearch"
+      expandable
+      size={props.iconSize}
+    >
+      <SearchBox.CollapsedState>
+        <SearchBox.ExpandButton
+          label="Search"
+          labelProps={{ placement: "bottom" }}
+          size={props.iconSize}
+          styleType="borderless"
+        >
+          <SvgSearch />
+        </SearchBox.ExpandButton>
+      </SearchBox.CollapsedState>
+      <SearchBox.ExpandedState>
+        <SearchBox.Icon size={props.iconSize}>
+          <SvgSearch />
+        </SearchBox.Icon>
+        <SearchBox.Input
+          ref={inputRef}
+          onChange={(e: React.ChangeEvent<HTMLInputElement>): void => {
+            setSearchText(e.currentTarget.value);
+            props.onSearch?.(e.currentTarget.value);
+          }}
+          value={searchText}
+        />
+        <SearchBox.CollapseButton
+          size={props.iconSize}
+          label="Clear"
+          labelProps={{ placement: "bottom" }}
+          onClick={() => {
+            setIsExpanded(false);
+            setSearchText("");
+            if (searchText) props.onSearch?.("");
+          }}
+        >
+          <SvgClose />
+        </SearchBox.CollapseButton>
+      </SearchBox.ExpandedState>
+    </SearchBox>
+  );
+}